requirements.txt
crewai>=0.28.0
langchain>=0.1.0
streamlit>=1.28.0
supabase>=2.3.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.2
asyncio
aiohttp